reorganized certain elements to avoid redundancy

+++ src/main.c
@@ -8,9 +8,7 @@
 #include <SDL3/SDL.h>
 #include <SDL3/SDL_main.h>
 #include <SDL3_ttf/SDL_ttf.h>
-#include <string.h>
 #include <stdbool.h>
-#include <time.h>
 #include "calculation_functions.h"
 #include "sdl_elements.h"
 
@@ -73,7 +71,7 @@ int main(int argc, char* argv[]) {
     // SDL ttf font stuff
     TTF_Init();
     g_font = TTF_OpenFont("CascadiaCode.ttf", wp.font_size);
-    g_font_small = TTF_OpenFont("CascadiaCode.ttf", wp.window_size_x / 90);
+    g_font_small = TTF_OpenFont("CascadiaCode.ttf", (float)wp.window_size_x / 90);
 
     ////////////////////////////////////////////////////////
     // simulation loop                                    //
